requirements.txt
# Data processing
pandas
numpy
openpyxl

# API requests
requests

# Testing
pytest

# Google Gemini API
google-genai
python-dotenv
